refactor: move EndTarget class to new end_target.py
