+++ gmatpyplus/command/__init__.py
@@ -1,5 +1,6 @@
 from .branch_command import *
 from .command import *
+from .end_target import *
 from .gmat_command import GmatCommand
 from .propagate import Propagate
 from .solver_branch_command import *
