+++ gmatpyplus/command/command.py
@@ -173,20 +173,6 @@ class EndFiniteBurn(GmatCommand):
         self.Initialize()
 
 
-class EndTarget(BranchCommand):
-    def __init__(self, parent_target: gp.Target, name: str = None):
-        if name is None:
-            name = f'EndTarget_{parent_target.GetName()}'
-
-        super().__init__('EndTarget', name)
-
-    def Insert(self, command: gp.GmatCommand | gmat.GmatCommand,
-               prev: gp.GmatCommand | gmat.GmatCommand = None) -> bool:
-        command = gp.extract_gmat_obj(command)
-        prev = gp.extract_gmat_obj(prev) if prev is not None else None
-        return gp.extract_gmat_obj(self).Insert(command, prev)
-
-
 class Maneuver(GmatCommand):
     def __init__(self, name: str, burn: gp.ImpulsiveBurn | gp.FiniteBurn, spacecraft: gp.Spacecraft,
                  backprop: bool = False):
